MD5/SHA1: use size_t for lengths

+++ src/crypto-md5.h
@@ -32,8 +32,7 @@ typedef struct MD5_CTX {
 } MD5_CTX;
 
 void MD5Init(MD5_CTX * context);
-void MD5Update(MD5_CTX * context, const unsigned char * data,
-               unsigned int len);
+void MD5Update(MD5_CTX * context, const unsigned char * data, size_t len);
 void MD5Final(unsigned char digest[16], MD5_CTX *context);
 
 #endif
